Refactor metrics
