+++ minivan/metrics.py
@@ -1,0 +1,17 @@
+import numpy as np
+
+DOT_PRODUCT = "dot_product"
+COSINE = "cosine"
+
+
+def normalize(embedding: np.ndarray) -> np.ndarray:
+    return embedding / np.linalg.norm(embedding, axis=-1, keepdims=True)
+
+
+def dot_product_metric(query_embedding: np.ndarray, embeddings: np.ndarray) -> np.ndarray:
+    return embeddings @ query_embedding
+
+
+def cosine_metric(query_embedding: np.ndarray, embeddings: np.ndarray) -> np.ndarray:
+    # assumes that embeddings are already normalized
+    return embeddings @ normalize(query_embedding)
