+++ tests/test_index.py
@@ -2,19 +2,6 @@ import numpy as np
 import pytest
 
 from minivan import Index
-from minivan.index import normalize
-
-
-def test_normalize():
-    embedding = np.array([0.1, 0.2, 0.3])
-    normalized_embedding = normalize(embedding)
-    assert np.allclose(normalized_embedding, embedding / np.linalg.norm(embedding))
-    assert np.allclose(np.power(normalized_embedding, 2).sum(), 1)
-
-    embedding = np.array([[0.1, 0.2, 0.3], [0.2, 0.3, 0.4]])
-    normalized_embedding = normalize(embedding)
-    assert np.allclose(normalized_embedding, embedding / np.linalg.norm(embedding, axis=1, keepdims=True))
-    assert np.allclose(np.power(normalized_embedding, 2).sum(axis=1), 1)
 
 
 def test_add_items_list():
